+++ src/main/java/cn/nukkit/inventory/PlayerInventory.java
@@ -360,7 +360,7 @@ public class PlayerInventory extends BaseInventory {
             if (player.equals(this.getHolder())) {
                 ContainerSetSlotPacket pk2 = new ContainerSetSlotPacket();
                 pk2.windowid = ContainerSetContentPacket.SPECIAL_ARMOR;
-                pk2.slot = (short) (index - this.getSize());
+                pk2.slot = index - this.getSize();
                 pk2.item = this.getItem(index);
                 player.dataPacket(pk2);
             } else {
@@ -431,7 +431,7 @@ public class PlayerInventory extends BaseInventory {
     @Override
     public void sendSlot(int index, Player[] players) {
         ContainerSetSlotPacket pk = new ContainerSetSlotPacket();
-        pk.slot = (short) index;
+        pk.slot = index;
         pk.item = this.getItem(index).clone();
 
         for (Player player : players) {
