+++ src/main/java/cn/nukkit/raknet/server/SessionManager.java
@@ -210,7 +210,7 @@ public class SessionManager {
                 RakNet.PACKET_RAW,
                 new byte[]{(byte) (address.length() & 0xff)},
                 address.getBytes(StandardCharsets.UTF_8),
-                Binary.writeShort((short) port),
+                Binary.writeShort(port),
                 payload
         );
         this.server.pushThreadToMainPacket(buffer);
@@ -244,7 +244,7 @@ public class SessionManager {
                 identifier.getBytes(StandardCharsets.UTF_8),
                 new byte[]{(byte) (session.getAddress().length() & 0xff)},
                 session.getAddress().getBytes(StandardCharsets.UTF_8),
-                Binary.writeShort((short) session.getPort()),
+                Binary.writeShort(session.getPort()),
                 Binary.writeLong(session.getID())
         );
         this.server.pushThreadToMainPacket(buffer);
