+++ src/main/java/cn/nukkit/tile/Furnace.java
@@ -34,16 +34,16 @@ public class Furnace extends Tile implements InventoryHolder, Container, Nameabl
         }
 
         if (!this.namedTag.contains("BurnTime") || this.namedTag.getShort("BurnTime") < 0) {
-            this.namedTag.putShort("BurnTime", (short) 0);
+            this.namedTag.putShort("BurnTime", 0);
         }
 
         if (!this.namedTag.contains("BurnTime") || this.namedTag.getShort("CookTime") < 0 || (this.namedTag.getShort("BurnTime") == 0 && this.namedTag.getShort("CookTime") > 0)) {
-            this.namedTag.putShort("CookTime", (short) 0);
+            this.namedTag.putShort("CookTime", 0);
         }
 
         if (!this.namedTag.contains("MaxTime")) {
             this.namedTag.putShort("MaxTime", this.namedTag.getShort("BurnTime"));
-            this.namedTag.putShort("BurnTicks", (short) 0);
+            this.namedTag.putShort("BurnTicks", 0);
         }
 
         if (this.namedTag.getShort("BurnTime") > 0) {
@@ -123,7 +123,7 @@ public class Furnace extends Tile implements InventoryHolder, Container, Nameabl
         CompoundTag d = new CompoundTag()
                 .putByte("Count", (byte) item.getCount())
                 .putByte("Slot", (byte) index)
-                .putShort("id", (short) item.getId())
+                .putShort("id", item.getId())
                 .putShort("Damage", item.getDamage());
 
         if (item.getId() == Item.AIR || item.getCount() <= 0) {
@@ -154,7 +154,7 @@ public class Furnace extends Tile implements InventoryHolder, Container, Nameabl
 
         this.namedTag.putShort("MaxTime", ev.getBurnTime());
         this.namedTag.putShort("BurnTime", ev.getBurnTime());
-        this.namedTag.putShort("BurnTicks", (short) 0);
+        this.namedTag.putShort("BurnTicks", 0);
         if (this.getBlock().getId() == Item.FURNACE) {
             this.getLevel().setBlock(this, Block.get(Item.BURNING_FURNACE, this.getBlock().getDamage()), true);
         }
@@ -186,11 +186,11 @@ public class Furnace extends Tile implements InventoryHolder, Container, Nameabl
         }
 
         if (this.namedTag.getShort("BurnTime") > 0) {
-            this.namedTag.putShort("BurnTime", (short) (this.namedTag.getShort("BurnTime") + 1));
-            this.namedTag.putShort("BurnTicks", (short) Math.ceil((double) this.namedTag.getShort("BurnTime") / (double) this.namedTag.getShort("MaxTime") * 200d));
+            this.namedTag.putShort("BurnTime", (this.namedTag.getShort("BurnTime") + 1));
+            this.namedTag.putShort("BurnTicks", (int) Math.ceil((double) this.namedTag.getShort("BurnTime") / (double) this.namedTag.getShort("MaxTime") * 200d));
 
             if (smelt != null && canSmelt) {
-                this.namedTag.putShort("CookTime", (short) (this.namedTag.getShort("CookTime") + 1));
+                this.namedTag.putShort("CookTime", (this.namedTag.getShort("CookTime") + 1));
                 if (this.namedTag.getShort("CookTime") >= 200) {
                     product = Item.get(smelt.getResult().getId(), smelt.getResult().getDamage(), product.getCount() + 1);
 
@@ -205,23 +205,23 @@ public class Furnace extends Tile implements InventoryHolder, Container, Nameabl
                         this.inventory.setSmelting(raw);
                     }
 
-                    this.namedTag.putShort("CookTime", (short) (this.namedTag.getShort("CookTime") - 200));
+                    this.namedTag.putShort("CookTime", (this.namedTag.getShort("CookTime") - 200));
                 }
             } else if (this.namedTag.getShort("BurnTime") <= 0) {
-                this.namedTag.putShort("BurnTime", (short) 0);
-                this.namedTag.putShort("CookTime", (short) 0);
-                this.namedTag.putShort("BurnTicks", (short) 0);
+                this.namedTag.putShort("BurnTime", 0);
+                this.namedTag.putShort("CookTime", 0);
+                this.namedTag.putShort("BurnTicks", 0);
             } else {
-                this.namedTag.putShort("CookTime", (short) 0);
+                this.namedTag.putShort("CookTime", 0);
             }
             ret = true;
         } else {
             if (this.getBlock().getId() == Item.BURNING_FURNACE) {
                 this.getLevel().setBlock(this, Block.get(Item.FURNACE, this.getBlock().getDamage()), true);
             }
-            this.namedTag.putShort("BurnTime", (short) 0);
-            this.namedTag.putShort("CookTime", (short) 0);
-            this.namedTag.putShort("BurnTicks", (short) 0);
+            this.namedTag.putShort("BurnTime", 0);
+            this.namedTag.putShort("CookTime", 0);
+            this.namedTag.putShort("BurnTicks", 0);
         }
 
         for (Player player : this.getInventory().getViewers()) {
